Tests/LibWeb/Text/expected/display_list/body-background-change-invalidates-root.txt

stale propagated background painted: false
fresh propagated background painted: true
